revert husky changes and example app tsx code refactor - added ListComponentScreen screen
