+++ src/components/floatingButton/index.ts
@@ -1,4 +1,4 @@
 import FloatingButton from './floatingButton';
-import floatingButtonStyle from './floatingButton.style';
+import { floatingButtonStyle } from './floatingButton.style';
 
 export { FloatingButton, floatingButtonStyle };
