+++ src/components/neoPopButton/index.ts
@@ -1,4 +1,4 @@
 import NeoPopButton from './neoPopButton';
-import neoPopButtonStyle from './neoPopButton.style';
+import { neoPopButtonStyle } from './neoPopButton.style';
 
 export { NeoPopButton, neoPopButtonStyle };
